[Bash] Remove unnecessary `sleep 1` from the check
Previously, `bash -c ls` inside Graphene resulted in a detached child
`ls` process (because execve emulation created an intermediate process
that silently exited, confusing make). To alleviate this issue, we
introduced a workaround of `sleep 1` to wait for the child to exit.
With new changes to Graphene, the intermediate process sticks around,
and make correctly waits for the child to finish everything.